Remove error message($!)
Because it is depend on locale. If user set locale to Japanese,
'$!' may be "接続を拒否されました", not "Connection refused".
And it is also depend platforms. This test expects connect system call
returns ECONNREFUSED, but FreeBSD connect system call returns EINVAL
(error message is "Invalid argument"). Windows may return other error.
So I remove error message and fix that it checks only that a error has occured.